only make a symlink to the app dir in our home directory if it's not already there

+++ user-config.sh
@@ -15,14 +15,13 @@ cp /vagrant/configs/bashrc /home/vagrant/.bashrc
 source /home/vagrant/synchro_venv/bin/activate
 # install pip packages
 cd /vagrant/synchro_app && pip install -r requirements.txt
-# install setuptools
-wget https://bootstrap.pypa.io/ez_setup.py -O - | sudo python
-rm setuptools*.zip
 
 # install npm packages
 cd /vagrant && npm install
 cd /vagrant/synchro_app/synchro/resources/bootstrap && npm install
 
-ln -s /vagrant/synchro_app ~/synchro_app
+if [ ! -L ~/synchro_app ]; then
+  ln -s /vagrant/synchro_app ~/synchro_app
+fi
